requirements.txt
opencv-contrib-python  # основной инструмент захвата и обработки кадров
pytesseract  # базовое решение задачи распознавание
tesseract  # дублирование пакета
pillow  # вывод результатов отображения в видеопоток
numpy  # часть функционала обрезки кадра
imutils  # часть фнкционала обрезки кадра